move FileInfoHeaderNoLookups to separate tarheader package
This aligns with how containerd organizes the code, and allows
consuming this function without requiring all the other code.
